Add files via upload
This project contains a program code for managing an emergency room using a priority queue. It will focus on the possible use of priority queueing during an emergency. Based on the severity of the cases, Hospitals distinguish the severity of cases and divide them into five general categories: Code, Critical, Urgent, Non-Urgent disabled, Ambulatory (most urgent to least urgent).